applications/video_converter/requirements.txt
# aws
boto3==1.34.55
# logging
loguru==0.7.2
# for env variables
python-dotenv==1.0.1